requirements_2004.txt
absl-py==2.1.0
alembic==1.13.1
astunparse==1.6.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
chex==0.1.86
clu==0.0.12
colorlog==6.8.2
-e git+ssh://git@github.com/nathan-t4/compgnn.git@48a7bdac47e624ab7166b8e10248c8903b7e3f64#egg=compgnn
contextlib2==21.6.0
contourpy==1.2.1
cycler==0.12.1
diffeqpy==2.4.1
diffrax==0.5.1
equinox==0.11.4
etils==1.8.0
fire==0.5.0
flatbuffers==24.3.25
flax==0.8.3
fonttools==4.51.0
fsspec==2024.5.0
gast==0.5.4
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.0
h5py==3.11.0
idna==3.7
importlib_resources==6.4.0
jax==0.4.28
jax-cuda12-pjrt==0.4.28
jax-cuda12-plugin==0.4.28
jaxlib==0.4.28
jaxtyping==0.2.28
jill==0.11.5
joblib==1.4.2
jraph==0.0.6.dev0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
juliacall==0.9.20
juliapkg==0.1.13
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
lineax==0.0.5
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.4.0
ml_collections==0.1.1
msgpack==1.0.8
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.4
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-cupti-cu12==12.3.101
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu12==12.3.107
nvidia-cuda-runtime-cu12==12.3.101
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu12==11.0.12.1
nvidia-curand-cu12==10.3.4.107
nvidia-cusolver-cu12==11.5.4.101
nvidia-cusparse-cu12==12.2.0.103
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
opt-einsum==3.3.0
optax==0.2.2
optimistix==0.0.7
optree==0.11.0
optuna==3.6.1
orbax-checkpoint==0.5.11
packaging==24.0
pillow==10.3.0
plotly==5.22.0
protobuf==4.25.3
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-gnupg==0.5.2
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.1
requests-futures==1.0.1
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.5.0
scipy==1.13.0
semantic-version==2.10.0
semver==3.0.2
six==1.16.0
SQLAlchemy==2.0.30
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
tensorstore==0.1.59
termcolor==2.4.0
threadpoolctl==3.5.0
toolz==0.12.1
tqdm==4.66.4
typeguard==2.13.3
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
wget==3.2
wrapt==1.16.0
zipp==3.18.2
